+++ ci/cpu/libcuspatial/build_libcuspatial.sh
@@ -1,5 +1,10 @@
 set -e
 
+# If nightly build, append current YYMMDD to version
+if [[ "$BUILD_MODE" = "branch" && "$SOURCE_BRANCH" = branch-* ]] ; then
+  export VERSION_SUFFIX=`date +%y%m%d`
+fi
+
 echo "Building libcuspatial"
 conda build conda/recipes/libcuspatial