+++ CHANGELOG.md
@@ -17,6 +17,8 @@
 - PR #44 Update all demos with Python API
 - PR #45 Improve documentation in haversine and point in polygon
 - PR #50 Validate that distance and speed work with all datetimes
+- PR #61 Point-in-polygon DataFrame output
+- PR #64 Use YYMMDD tag in nightly build
 
 ## Bug Fixes
 
@@ -24,3 +26,6 @@
 - PR #17 Update issue / PR templates
 - PR #23 Fix cudf Cython imports
 - PR #24 `cuspatial::derive_trajectories()` test improvements and bug fixes
+- PR #33 `cuspatial::trajectory_distance_and_speed()` test improvements and bug fixes
+- PR #49 Docstring for haversine and argument ordering was backwards
+- PR #66 added missing header in tests