+++ CHANGELOG.md
@@ -16,8 +16,10 @@
 - PR #41 Python and Cython style cleanup, pre-commit hook
 - PR #44 Update all demos with Python API
 - PR #45 Improve documentation in haversine and point in polygon
+- PR #50 Validate that distance and speed work with all datetimes
 - PR #61 Point-in-polygon DataFrame output
 - PR #59 Improve detail of point in polygon docs
+- PR #64 Use YYMMDD tag in nightly build
 
 ## Bug Fixes
 
@@ -25,4 +27,6 @@
 - PR #17 Update issue / PR templates
 - PR #23 Fix cudf Cython imports
 - PR #24 `cuspatial::derive_trajectories()` test improvements and bug fixes
+- PR #33 `cuspatial::trajectory_distance_and_speed()` test improvements and bug fixes
 - PR #49 Docstring for haversine and argument ordering was backwards
+- PR #66 added missing header in tests