+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - PR #45 Improve documentation in haversine and point in polygon
 - PR #50 Validate that distance and speed work with all datetimes
 - PR #61 Point-in-polygon DataFrame output
+- PR #59 Improve detail of point in polygon docs
 - PR #64 Use YYMMDD tag in nightly build
 
 ## Bug Fixes