+++ CHANGELOG.md
@@ -16,7 +16,9 @@
 - PR #41 Python and Cython style cleanup, pre-commit hook
 - PR #44 Update all demos with Python API
 - PR #45 Improve documentation in haversine and point in polygon
+- PR #50 Validate that distance and speed work with all datetimes
 - PR #61 Point-in-polygon DataFrame output
+- PR #59 Improve detail of point in polygon docs
 - PR #64 Use YYMMDD tag in nightly build
 
 ## Bug Fixes