+++ CHANGELOG.md
@@ -19,7 +19,9 @@
 - PR #50 Validate that distance and speed work with all datetimes
 - PR #58 Hausdorff distance returns a DataFrame, and better docs.
 - PR #61 Point-in-polygon DataFrame output
+- PR #59 Improve detail of point in polygon docs
 - PR #64 Use YYMMDD tag in nightly build
+- PR #68 Use YYMMDD tag in nightly build of cuspatial python 
 
 ## Bug Fixes