+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# cuSpatial 0.10.0 (Date TBD)
+# cuSpatial 0.10.0 (Oct 16 2019)
 
 ## New Features